etl-solution/processor/requirements.txt
psycopg2-binary==2.8.5
confluent-kafka==1.4.0
json-logging==1.1.0
retry==0.9.2